+++ src/main/factories/signup/signup-validation.spec.ts
@@ -1,5 +1,6 @@
-import { CompareFieldsValidation, EmailValidation, RequiredFieldValidation, Validation, ValidationComposite } from '../../../presentation/helpers/validators'
+import { CompareFieldsValidation, EmailValidation, RequiredFieldValidation, ValidationComposite } from '../../../presentation/helpers/validators'
 import { EmailValidator } from '../../../presentation/protocols/email-validator'
+import { Validation } from '../../../presentation/protocols/validation'
 import { makeSignUpValidation } from './signup-validation'
 
 jest.mock('../../../presentation/helpers/validators/validation-composite')
