+++ src/main/factories/signup/signup-validation.ts
@@ -1,4 +1,5 @@
-import { CompareFieldsValidation, EmailValidation, RequiredFieldValidation, Validation, ValidationComposite } from '../../../presentation/helpers/validators'
+import { CompareFieldsValidation, EmailValidation, RequiredFieldValidation, ValidationComposite } from '../../../presentation/helpers/validators'
+import { Validation } from '../../../presentation/protocols/validation'
 import { EmailValidatorAdapter } from '../../../utils/email-validator-adapter'
 
 export const makeSignUpValidation = (): ValidationComposite => {
