+++ src/presentation/controllers/login/login-protocols.ts
@@ -2,4 +2,4 @@
 export * from '../../protocols'
 export * from '../../protocols/email-validator'
 export * from '../../../domain/usecases/authentication'
-export * from '../../helpers/validators/validation'
+export * from '../../protocols/validation'
