+++ src/presentation/controllers/signup/signup-protocols.ts
@@ -3,4 +3,4 @@ export * from '../../protocols'
 export * from '../../protocols/email-validator'
 export * from '../../../domain/usecases/add-account'
 export * from '../../../domain/models/account'
-export * from '../../helpers/validators/validation'
+export * from '../../protocols/validation'
