+++ src/presentation/helpers/validators/compare-fields-validation.ts
@@ -1,5 +1,5 @@
 import { InvalidParamError } from '../../erros'
-import { Validation } from './validation'
+import { Validation } from '../../protocols/validation'
 
 export class CompareFieldsValidation implements Validation {
   constructor (private readonly fieldName: string, private readonly fieldToCompareName: string) {}
