+++ src/presentation/helpers/validators/index.ts
@@ -2,5 +2,4 @@
 export * from './email-validation'
 export * from './compare-fields-validation'
 export * from './required-field-validation'
-export * from './validation'
 export * from './validation-composite'
